+++ modules/AWSPowerShell/Cmdlets/AmazonLightsail/Basic/New-LSInstance-Cmdlet.cs
@@ -80,7 +80,8 @@ namespace Amazon.PowerShell.Cmdlets.LS
         #region Parameter CustomImageName
         /// <summary>
         /// <para>
-        /// <para>The name for your custom image.</para>
+        /// <para>(Deprecated) The name for your custom image.</para><note><para>In releases prior to June 12, 2017, this parameter was ignored by the API. It is now
+        /// deprecated.</para></note>
         /// </para>
         /// </summary>
         [System.Management.Automation.Parameter]