+++ modules/AWSPowerShell/Cmdlets/AmazonRoute53/Basic/Edit-R53ResourceRecordSet-Cmdlet.cs
@@ -84,9 +84,8 @@ namespace Amazon.PowerShell.Cmdlets.R53
     /// your changes to all of the Amazon Route 53 authoritative DNS servers. While your changes
     /// are propagating, <code>GetChange</code> returns a status of <code>PENDING</code>.
     /// When propagation is complete, <code>GetChange</code> returns a status of <code>INSYNC</code>.
-    /// Changes generally propagate to all Amazon Route 53 name servers in a few minutes.
-    /// In rare circumstances, propagation can take up to 30 minutes. For more information,
-    /// see <a>GetChange</a>.
+    /// Changes generally propagate to all Amazon Route 53 name servers within 60 seconds.
+    /// For more information, see <a>GetChange</a>.
     /// </para><para><b>Limits on ChangeResourceRecordSets Requests</b></para><para>
     /// For information about the limits on a <code>ChangeResourceRecordSets</code> request,
     /// see <a href="http://docs.aws.amazon.com/Route53/latest/DeveloperGuide/DNSLimitations.html">Limits</a>